lambda_functions/requirements.txt
aiohttp==3.9.1
aiofiles==23.2.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
psycopg2-binary==2.9.7
boto3==1.29.7